+++ src/main/resources/config.properties
@@ -1,0 +1,2 @@
+app.url = https://www.criticker.com/
+chromedriver.path = drivers/chromedriver.exe
