+++ src/main/resources/features/Vote.feature
