--- src/test/java/constant/LoginPageConstant.java
@@ -1,9 +1,0 @@
-package constant;
-
-import org.openqa.selenium.By;
-
-public class LoginPageConstant {
-    
-	public static final By SI_INPUT_UNAME_CSS_SELECTOR = By.cssSelector("#si_input_uname");
-	public static final By SI_INPUT_PSWD_CSS_SELECTOR = By.cssSelector("#si_input_pswd");
-}
