--- src/test/resources/config.properties
@@ -1,2 +1,0 @@
-app.url = https://www.criticker.com/
-chromedriver.path = drivers/chromedriver.exe
